# packages/contracts/http/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Place HTTP",
    "version": "1.0.0",
    "description": "Place-owned source-only HTTP and browser BFF contract."
  },
  "servers": [],
  "paths": {
...
    "/api/imports": {
      "post": {
        "operationId": "requestPlaceImportForBrowser",
        "summary": "requestPlaceImportForBrowser",
        "security": [
          {
            "placeBrowserSession": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceImportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Return an idempotently replayed import batch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceImportBatch"
                }
              }
            }
          },
          "202": {
            "description": "Queue a connected-account import",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceImportBatch"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProductRequestInvalid"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "409": {
            "$ref": "#/components/responses/ProductConflict"
          },
          "503": {
            "$ref": "#/components/responses/BrowserBackendUnavailable"
          }
        }
      }
    },
    "/api/imports/{batchId}": {
      "parameters": [
        {
          "name": "batchId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getPlaceImportForBrowser",
        "summary": "getPlaceImportForBrowser",
        "security": [
          {
            "placeBrowserSession": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2048
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return a bounded import preview and progress",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceImportBatchDetail"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProductRequestInvalid"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "404": {
            "$ref": "#/components/responses/ProductNotFound"
          },
          "503": {
            "$ref": "#/components/responses/BrowserBackendUnavailable"
          }
        }
      }
    },
...
    "/api/public/collections/{publicationId}": {
      "parameters": [
        {
          "name": "publicationId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getPublishedPlaceCollectionForBrowser",
        "summary": "getPublishedPlaceCollectionForBrowser",
        "security": [],
        "responses": {
          "200": {
            "description": "Return a validated public collection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishedCollection"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ProductNotFound"
          },
          "503": {
            "$ref": "#/components/responses/BrowserBackendUnavailable"
          }
        }
      }
    },
...
    "/v1/library/commands": {
      "post": {
        "operationId": "applyPlaceLibraryCommand",
        "summary": "applyPlaceLibraryCommand",
        "security": [
          {
            "placeBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Return an idempotently replayed command result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryCommandResult"
                }
              }
            }
          },
          "201": {
            "description": "Return an applied command result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryCommandResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProductRequestInvalid"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "404": {
            "$ref": "#/components/responses/ProductNotFound"
          },
          "409": {
            "$ref": "#/components/responses/ProductConflict"
          }
        }
      }
    },
...
    "/v1/imports": {
      "get": {
        "operationId": "listCurrentMemberPlaceImports",
        "summary": "listCurrentMemberPlaceImports",
        "security": [
          {
            "placeBearer": []
          }
        ],
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "queued",
                "running",
                "partial",
                "enriching",
                "needs-user-action",
                "needs-review",
                "completed",
                "failed",
                "cancelled"
              ],
              "default": "all"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2048
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return a bounded current-member ImportBatch history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceImportBatchList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProductRequestInvalid"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "503": {
            "$ref": "#/components/responses/ImportQueryUnavailable"
          }
        }
      },
      "post": {
        "operationId": "requestPlaceImport",
        "summary": "requestPlaceImport",
        "security": [
          {
            "placeBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceImportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Return an idempotently replayed import batch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceImportBatch"
                }
              }
            }
          },
          "202": {
            "description": "Queue a connected-account import",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceImportBatch"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProductRequestInvalid"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "409": {
            "$ref": "#/components/responses/ProductConflict"
          }
        }
      }
    },
    "/v1/imports/{batchId}": {
      "parameters": [
        {
          "name": "batchId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getPlaceImport",
        "summary": "getPlaceImport",
        "security": [
          {
            "placeBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2048
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return an import preview and progress",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceImportBatchDetail"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProductRequestInvalid"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "404": {
            "$ref": "#/components/responses/ProductNotFound"
          },
          "503": {
            "$ref": "#/components/responses/ImportQueryUnavailable"
          }
        }
      }
    },
...
    "/v1/library/places/{placeId}": {
      "parameters": [
        {
          "name": "placeId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getPlacePreferences",
        "summary": "getPlacePreferences",
        "security": [
          {
            "placeBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Return current-member Place preferences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryPlacePreferences"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ProductRequestInvalid"
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationRequired"
          },
          "403": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "404": {
            "$ref": "#/components/responses/ProductNotFound"
          }
        }
      }
    },
...
    "/v1/public/collections/{publicationId}": {
      "parameters": [
        {
          "name": "publicationId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getPublishedPlaceCollection",
        "summary": "getPublishedPlaceCollection",
        "security": [],
        "responses": {
          "200": {
            "description": "Return an allowlisted public collection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishedCollection"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ProductNotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "placeBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "placeBrowserSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-place_session"
      },
      "placeConnector": {
        "type": "http",
        "scheme": "PlaceConnector"
      }
    },
    "schemas": {
...
      "LibraryCommandResult": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "library-command-result.v1"
          },
          "status": {
            "type": "string",
            "enum": [
              "applied",
              "replayed"
            ]
          }
        },
        "required": [
          "schemaVersion",
          "status"
        ],
        "additionalProperties": false
      },
      "LibraryPlacePreferences": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "library-place-preferences.v1"
          },
          "placeId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "saved": {
            "type": "boolean"
          },
          "wanted": {
            "type": "boolean"
          },
          "personalRating": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.1,
                "maximum": 5,
                "multipleOf": 0.1
              },
              {
                "type": "null"
              }
            ]
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          }
        },
        "required": [
          "schemaVersion",
          "placeId",
          "saved",
          "wanted",
          "personalRating",
          "updatedAt"
        ],
        "additionalProperties": false
      },
...
      "PlaceImportRequest": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "place-import-request.v1"
          },
          "connectionId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "idempotencyKey": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          }
        },
        "required": [
          "schemaVersion",
          "connectionId",
          "idempotencyKey"
        ],
        "additionalProperties": false
      },
      "PlaceImportBatch": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "place-import-batch.v1"
          },
          "batchId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "connectionId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "providerKey": {
            "type": "string",
            "enum": [
              "naver",
              "kakao",
              "google"
            ]
          },
          "state": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "partial",
              "enriching",
              "needs-user-action",
              "needs-review",
              "completed",
              "failed",
              "cancelled"
            ]
          },
          "progress": {
            "type": "object",
            "properties": {
              "discovered": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "ready": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "reviewRequired": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "enriching": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "applied": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "skipped": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "failed": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "discovered",
              "ready",
              "reviewRequired",
              "enriching",
              "applied",
              "skipped",
              "failed"
            ],
            "additionalProperties": false
          },
          "failure": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "provider-auth-expired",
                  "provider-mfa-required",
                  "provider-captcha-required",
                  "provider-consent-required",
                  "provider-rate-limited",
                  "provider-parser-drift",
                  "provider-unavailable",
                  "capture-invalid",
                  "internal-failure"
                ]
              },
              "retryable": {
                "type": "boolean"
              }
            },
            "required": [
              "code",
              "retryable"
            ],
            "additionalProperties": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          }
        },
        "required": [
          "schemaVersion",
          "batchId",
          "connectionId",
          "providerKey",
          "state",
          "progress",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
...
      "PublishedCollection": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "place-published-collection.v1"
          },
          "publicationId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "unlisted",
              "public"
            ]
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ]
          },
          "places": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "placeId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "position": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                }
              },
              "required": [
                "placeId",
                "position"
              ],
              "additionalProperties": false
            }
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          }
        },
        "required": [
          "schemaVersion",
          "publicationId",
          "visibility",
          "name",
          "description",
          "places",
          "updatedAt"
        ],
        "additionalProperties": false
      },
...
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "minLength": 1
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "code": {
            "type": "string",
            "minLength": 1
          },
          "retryable": {
            "type": "boolean"
          },
          "correlationRef": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "retryable",
          "correlationRef"
        ],
        "additionalProperties": false
      }
    },
    "responses": {
...
      "BrowserBackendUnavailable": {
        "description": "The fixed internal Place Backend is temporarily unavailable",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "ProductRequestInvalid": {
        "description": "The product request is malformed or contains unsupported fields",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "ProductNotFound": {
        "description": "The owned resource or disclosed publication is unavailable",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "ImportQueryUnavailable": {
        "description": "The bounded Import query is temporarily unavailable",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "ProductConflict": {
        "description": "The request conflicts with prior state",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "AuthenticationRequired": {
        "description": "The bearer token is missing or invalid",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "AccessDenied": {
        "description": "The principal lacks Place access",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
    }
  }
}